Apply code reviewer revisions
