+++ src/elements.js
@@ -1,0 +1,5 @@
+const postContainer = document.querySelector('.tasklist');
+const addButton = document.querySelector('.newtask-return');
+const addNewInput = document.getElementById('addnew');
+
+export { postContainer, addButton, addNewInput };
